requirements.txt
langchain 
langgraph 
langchain_community 
langchain_experimental 
langchain-groq 
langchain-google-genai
arxiv 
google-search-results 
python-docx
python-dotenv
duckduckgo-search
agno
pypdf
google-genai
gspread
oauth2client
replicate
aiohttp
aiofiles
openai
google-api-python-client 
google-auth-oauthlib 
google-auth-httplib2